rename chat module
